Updates: little cleanup, separate dev mode code
Trying to add more comments for documentation and clean up other
comments in the code. Still WIP.

Separating the device mode code into it's own file. This is to make
authoring that part of the code a little easier.
Some of the RTC code needed to be moved as a result.